requirements.txt
# OpenKitx403 Bot API Requirements
fastapi>=0.104.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
base58>=2.1.1
pynacl>=1.5.0
solders>=0.18.1

